Minuto passa a admitir fração: planejada x OEE volta a fechar
Ajuste 6: 29.430 min com OEE de 75% dava 22.092 e não 22.072,5. Não era a
fórmula: o motor fazia round(planejada x oee) em CADA linha — o grão é recurso
x dia x turno — e o mês era a soma dessas linhas já arredondadas. Meio minuto
sobrando numa, um quarto noutra, e no fim do mês a soma não bate com a
multiplicação. Pior: duas máquinas com o mesmo total mensal podiam dar
disponíveis diferentes, dependendo de como os minutos se dividiam nos dias.

O round sai. min_instalada, min_planejada, min_disponivel e as três colunas do
memorial viram numeric(18,6), e arredondar passa a ser decisão da tela — uma
vez só, no fim. Em minutos a tela mostra a casa decimal quando ela existe, que
é justamente o número conferido na calculadora.

Ajuste 7: Código antes de Recurso nos três seletores de planejamento, na coluna
da tabela de paradas e no cabeçalho dos painéis.

Migração: 12_minuto_fracao.sql, antes do deploy. Depois, Recalcular — rodadas
antigas continuam com os números antigos.

+++ 01_schema.sql
@@ -412,7 +412,9 @@ create table capacidade_instalada_dia (
     data          date   not null,
     qt_recursos   int    not null,
     equivalencia  numeric(10,4) not null,
-    min_instalada bigint not null,
+    -- Numeric e nao bigint: arredondar minuto no meio da cadeia faz a soma
+    -- do mes discordar da multiplicacao. Ver 12_minuto_fracao.sql.
+    min_instalada numeric(18,6) not null,
     primary key (execucao_id, recurso_id, data)
 );
 create index ix_cid_area_data on capacidade_instalada_dia (area_id, data);
@@ -438,8 +440,8 @@ create table capacidade_fato (
     min_turno_liquido    int    not null,
     min_parada_planejada int    not null default 0,
     min_parada_outras    int    not null default 0,
-    min_planejada        bigint not null,
-    min_disponivel       bigint not null,
+    min_planejada        numeric(18,6) not null,
+    min_disponivel       numeric(18,6) not null,
 
     unidade_medida_id    int references unidade_medida(id),
     qtd_planejada        numeric(18,4),
@@ -459,9 +461,9 @@ create table capacidade_memoria (
     turno_id       int  not null,
     ordem          smallint not null,
     etapa          varchar(40) not null,
-    minutos_antes  bigint not null,
-    minutos_delta  bigint not null,
-    minutos_depois bigint not null,
+    minutos_antes  numeric(18,6) not null,
+    minutos_delta  numeric(18,6) not null,
+    minutos_depois numeric(18,6) not null,
     origem_tabela  varchar(40),
     origem_id      bigint,
     descricao      varchar(200)
